examples/marketing-assets/backend/pyproject.toml
[project]
name = "chatkit-backend"
version = "0.1.0"
description = "FastAPI backend service for ChatKit boilerplate"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.114.1,<0.116",
    "httpx>=0.28,<0.29",
    "uvicorn[standard]>=0.30,<0.31",
    "openai>=1.40",
    "openai-chatkit>=0.0.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.4,<0.7",
    "mypy>=1.8,<2",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = ["I"]

# TODO: remove this when making public
[tool.uv.sources]
chatkit = { path = "../../../../chatkit-python-internal" }
